restructure deserialisation
remove code duplication and simplify lambda expression
