+++ src/serialiser/include/IoSerialiser.hpp
@@ -11,6 +11,50 @@
 #pragma ide diagnostic   ignored "cppcoreguidelines-avoid-magic-numbers"
 #pragma ide diagnostic   ignored "cppcoreguidelines-avoid-c-arrays"
 
+struct START_MARKER {};
+template<>
+struct fmt::formatter<START_MARKER> {
+    template<typename ParseContext>
+    constexpr auto parse(ParseContext &ctx) {
+        auto it = ctx.begin(), end = ctx.end();
+        // Check if reached the end of the range:
+        if (it != end && *it != '}')
+            throw format_error("invalid format");
+        // Return an iterator past the end of the parsed range:
+        return it;
+    }
+
+    template<typename FormatContext>
+    auto format(START_MARKER const &startMarker, FormatContext &ctx) {
+        std::ignore = startMarker;
+        return fmt::format_to(ctx.out(), "START_MARKER\n");
+    }
+};
+
+struct END_MARKER {};
+template<>
+struct fmt::formatter<END_MARKER> {
+    template<typename ParseContext>
+    constexpr auto parse(ParseContext &ctx) {
+        auto it = ctx.begin(), end = ctx.end();
+        // Check if reached the end of the range:
+        if (it != end && *it != '}')
+            throw format_error("invalid format");
+        // Return an iterator past the end of the parsed range:
+        return it;
+    }
+    template<typename FormatContext>
+    auto format(END_MARKER const &endMarker, FormatContext &ctx) {
+        std::ignore = endMarker;
+        return fmt::format_to(ctx.out(), "END_MARKER\n");
+    }
+};
+
+struct OTHER {};
+
+constexpr static START_MARKER START_MARKER_INST;
+constexpr static END_MARKER   END_MARKER_INST;
+
 namespace opencmw {
 
 enum ProtocolCheck {
@@ -86,544 +130,336 @@ struct IoSerialiser {
     }
 };
 
-} // namespace opencmw
-
-/* #################################################################################### */
-/* ### Yet-Another-Serialiser - YaS Definitions ####################################### */
-/* #################################################################################### */
-struct START_MARKER {};
-template<>
-struct fmt::formatter<START_MARKER> {
-    template<typename ParseContext>
-    constexpr auto parse(ParseContext &ctx) {
-        auto it = ctx.begin(), end = ctx.end();
-        // Check if reached the end of the range:
-        if (it != end && *it != '}')
-            throw format_error("invalid format");
-        // Return an iterator past the end of the parsed range:
-        return it;
-    }
-
-    template<typename FormatContext>
-    auto format(START_MARKER const &startMarker, FormatContext &ctx) {
-        std::ignore = startMarker;
-        return fmt::format_to(ctx.out(), "START_MARKER\n");
-    }
+struct FieldDescription {
+    uint64_t         headerStart;
+    uint8_t          intDataType;
+    uint64_t         hash;
+    uint64_t         dataStartOffset;
+    uint64_t         dataSize;
+    std::string_view fieldName;
+    std::size_t      dataStartPosition;
+    std::size_t      dataEndPosition;
+    std::string_view unit;
+    std::string_view description;
+    ExternalModifier modifier;
 };
 
-struct END_MARKER {};
-template<>
-struct fmt::formatter<END_MARKER> {
-    template<typename ParseContext>
-    constexpr auto parse(ParseContext &ctx) {
-        auto it = ctx.begin(), end = ctx.end();
-        // Check if reached the end of the range:
-        if (it != end && *it != '}')
-            throw format_error("invalid format");
-        // Return an iterator past the end of the parsed range:
-        return it;
-    }
-    template<typename FormatContext>
-    auto format(END_MARKER const &endMarker, FormatContext &ctx) {
-        std::ignore = endMarker;
-        return fmt::format_to(ctx.out(), "END_MARKER\n");
-    }
-};
-
-struct OTHER {};
-
-constexpr static START_MARKER START_MARKER_INST;
-constexpr static END_MARKER   END_MARKER_INST;
-
-namespace opencmw {
-struct YaS : Protocol<"YaS"> {};
-
-namespace yas {
-static const int              VERSION_MAGIC_NUMBER = -1;    // '-1' since CmwLight cannot have a negative number of entries
-static const std::string_view PROTOCOL_NAME        = "YaS"; // Yet another Serialiser implementation
-static const uint8_t          VERSION_MAJOR        = 1;
-static const uint8_t          VERSION_MINOR        = 0;
-static const uint8_t          VERSION_MICRO        = 0;
-static const uint8_t          ARRAY_TYPE_OFFSET    = 100U;
-
-// clang-format off
-template<typename T> inline constexpr uint8_t getDataTypeId() { return 0xFF; } // default value
-template<> inline constexpr uint8_t getDataTypeId<START_MARKER>() { return 0; }
-template<> inline constexpr uint8_t getDataTypeId<bool>() { return 1; }
-template<> inline constexpr uint8_t getDataTypeId<int8_t>() { return 2; }
-template<> inline constexpr uint8_t getDataTypeId<int16_t>() { return 3; }
-template<> inline constexpr uint8_t getDataTypeId<int32_t>() { return 4; }
-template<> inline constexpr uint8_t getDataTypeId<int64_t>() { return 5; }
-template<> inline constexpr uint8_t getDataTypeId<float>() { return 6; }
-template<> inline constexpr uint8_t getDataTypeId<double>() { return 7; }
-template<> inline constexpr uint8_t getDataTypeId<char>() { return 8; }
-template<> inline constexpr uint8_t getDataTypeId<std::string>() { return 9; }
-
-template<> inline constexpr uint8_t getDataTypeId<bool[]>() { return 101; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<bool>>() { return 101; }
-template<> inline constexpr uint8_t getDataTypeId<int8_t[]>() { return 102; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<int8_t>>() { return 102; }
-template<> inline constexpr uint8_t getDataTypeId<int16_t[]>() { return 103; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<int16_t>>() { return 103; }
-template<> inline constexpr uint8_t getDataTypeId<int32_t[]>() { return 104; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<int32_t>>() { return 104; }
-template<> inline constexpr uint8_t getDataTypeId<int64_t[]>() { return 105; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<int64_t>>() { return 105; }
-template<> inline constexpr uint8_t getDataTypeId<float[]>() { return 106; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<float>>() { return 106; }
-template<> inline constexpr uint8_t getDataTypeId<double[]>() { return 107; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<double>>() { return 107; }
-template<> inline constexpr uint8_t getDataTypeId<char[]>() { return 108; }
-template<> inline constexpr uint8_t getDataTypeId<std::vector<char>>() { return 108; }
-
-template<> inline constexpr uint8_t getDataTypeId<std::string[]>() { return 109; }
-template<> inline constexpr uint8_t getDataTypeId<std::string_view[]>() { return 109; }
-
-// template<> inline constexpr uint8_t getDataTypeId<START_MARKER>() { return 200; }
-// template<> inline constexpr uint8_t getDataTypeId<enum>()          { return 201; }
-// template<typename T> inline constexpr uint8_t getDataTypeId<std::list<T>>()     { return 202; }
-// template<> inline constexpr uint8_t getDataTypeId<std::unsorted_map>() { return 203; }
-// template<> inline constexpr uint8_t getDataTypeId<std::queue>()    { return 204; }
-// template<> inline constexpr uint8_t getDataTypeId<std::set>()      { return 205; }
-
-template<>
-inline constexpr uint8_t getDataTypeId<OTHER>() { return 0xFD; }
-
-template<>
-inline constexpr uint8_t getDataTypeId<END_MARKER>() { return 0xFE; }
-
-// clang-format on
-} // namespace yas
-
-template<typename T>
-struct IoSerialiser<YaS, T> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<T>(); } // default value
-};
-
-template<Number T> // catches all numbers
-struct IoSerialiser<YaS, T> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<T>(); }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        buffer.put(value);
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) noexcept {
-        value = buffer.get<T>();
-        return std::is_constant_evaluated();
-    }
-};
-
-template<StringLike T>
-struct IoSerialiser<YaS, T> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<T>(); }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        //        std::cout << fmt::format("{} - serialise-String_like: {} {} == {} - constexpr?: {} - const {}\n",
-        //                YaS::protocolName(), typeName<T>(), field, value, std::is_constant_evaluated(), std::is_const_v<T>);
-        buffer.put<T>(value); // N.B. ensure that the wrapped value and not the annotation itself is serialised
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) noexcept {
-        value = buffer.get<std::string>();
-        //        std::cout << fmt::format("{} - de-serialise-String_like: {} {} == {}  {} - constexpr?: {}\n",
-        //            YaS::protocolName(), typeName<T>(), field, value, value.size(), std::is_constant_evaluated());
-        return std::is_constant_evaluated();
-    }
-};
-
-template<ArrayOrVector T>
-struct IoSerialiser<YaS, T> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<T>(); }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        buffer.put(std::array<int32_t, 1>{ static_cast<int32_t>(value.size()) });
-        buffer.put(value);
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) noexcept {
-        buffer.getArray<int32_t, 1>();
-        buffer.getArray(value);
-        return std::is_constant_evaluated();
-    }
-};
-
-template<MultiArrayType T>
-struct IoSerialiser<YaS, T> {
-    inline static constexpr uint8_t getDataTypeId() {
-        // std::cout << fmt::format("getDataTypeID<{}>() = {}\n", typeName<typename T::value_type>(), yas::getDataTypeId<typename T::value_type>());
-        return yas::ARRAY_TYPE_OFFSET + yas::getDataTypeId<typename T::value_type>();
-    }
-    constexpr static bool serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        std::array<int32_t, T::n_dims_> dims;
-        for (uint32_t i = 0U; i < T::n_dims_; i++) {
-            dims[i] = static_cast<int32_t>(value.dimensions()[i]);
+/**
+ * Reflection helper, calls the given lambda expresion
+ * @tparam T Type parameter for the object
+ * @tparam Callable Type parameter for the callable
+ * @param fieldName The name of the field to call the callable with
+ * @param callable A lambda expression refl::member, index -> void which will be called for field with the given name
+ */
+// TODO move to refl, change function name corresponding to ranges api
+template<ReflectableClass T, typename Callable>
+void call_on_field(const std::string_view fieldName, const Callable &callable) {
+    const auto searchIndex = static_cast<int32_t>(findMemberIndex<T>(fieldName));
+    for_each(refl::reflect<T>().members, [&callable, &searchIndex](auto member, int32_t index) {
+        if (index != searchIndex) {
+            return; // fieldName does not match -- skip to next field
         }
-        buffer.put(dims);
-        buffer.put(value.elements()); // todo: account for strides and offsets (possibly use iterators?)
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) noexcept {
-        const std::array<int32_t, T::n_dims_> dimWire = buffer.getArray<int32_t, T::n_dims_>();
-        for (auto i = 0U; i < T::n_dims_; i++) {
-            value.dimensions()[i] = static_cast<typename T::size_t_>(dimWire[i]);
-        }
-        value.element_count()        = value.dimensions()[T::n_dims_];
-        value.stride(T::n_dims_ - 1) = 1;
-        value.offset(T::n_dims_ - 1) = 0;
-        for (auto i = T::n_dims_ - 1; i > 0; i--) {
-            value.element_count() *= value.dimensions()[i - 1];
-            value.stride(i - 1) = value.stride(i) * value.dimensions()[i];
-            value.offset(i - 1) = 0;
-        }
-        buffer.getArray(value.elements());
-        return std::is_constant_evaluated();
-    }
-};
-
-template<>
-struct IoSerialiser<YaS, START_MARKER> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<START_MARKER>(); }
-
-    constexpr static bool           serialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const START_MARKER & /*value*/) noexcept {
-        // do not do anything, as the start marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-
-    constexpr static bool deserialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const START_MARKER &) {
-        // do not do anything, as the start marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-};
-
-template<>
-struct IoSerialiser<YaS, END_MARKER> {
-    inline static constexpr uint8_t getDataTypeId() { return yas::getDataTypeId<END_MARKER>(); }
-    static bool                     serialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const END_MARKER & /*value*/) noexcept {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-
-    constexpr static bool deserialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const END_MARKER &) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-};
-
-template<>
-struct FieldHeader<YaS> {
-    template<const bool writeMetaInfo, typename DataType>
-    constexpr std::size_t static putFieldHeader(IoBuffer &buffer, const char *fieldName, const int fieldNameSize, const DataType &data) { // todo fieldName -> string_view
-        using StrippedDataType         = std::remove_reference_t<decltype(getAnnotatedMember(unwrapPointer(data)))>;
-        constexpr int32_t dataTypeSize = static_cast<int32_t>(sizeof(StrippedDataType));
-        buffer.reserve_spare(((static_cast<uint64_t>(fieldNameSize) + 18) * sizeof(uint8_t)) + dataTypeSize);
-
-        // -- offset 0 vs. field start
-        const std::size_t headerStart = buffer.size();
-        buffer.put(static_cast<uint8_t>(IoSerialiser<YaS, StrippedDataType>::getDataTypeId())); // data type ID
-        buffer.put(opencmw::hash(fieldName, fieldNameSize));                                    // unique hashCode identifier -- TODO: choose more performant implementation instead of java default
-        const std::size_t dataStartOffsetPosition = buffer.size();
-        buffer.put(-1); // dataStart offset
-        const int32_t     dataSize         = is_supported_number<StrippedDataType> ? dataTypeSize : -1;
-        const std::size_t dataSizePosition = buffer.size();
-        buffer.put(dataSize);                    // dataSize (N.B. 'headerStart' + 'dataStart + dataSize' == start of next field header
-        buffer.put<std::string_view>(fieldName); // full field name
-
-        if constexpr (is_annotated<DataType> && writeMetaInfo) {
-            //if (writeMetaInfo) {
-            buffer.put(std::string_view(data.getUnit()));
-            buffer.put(std::string_view(data.getDescription()));
-            buffer.put(static_cast<uint8_t>(data.getModifier()));
-            // TODO: write group meta data
-            // final String[] groups = fieldDescription.getFieldGroups().toArray(new String[0]); // java uses non-array string
-            // buffer.putStringArray(groups, groups.length);
-            // buffer.put<std::string[]>({""});
-            //}
-        }
-
-        const std::size_t dataStartPosition         = buffer.size();
-        const std::size_t dataStartOffset           = (dataStartPosition - headerStart);     // -- offset dataStart calculations
-        buffer.at<int32_t>(dataStartOffsetPosition) = static_cast<int32_t>(dataStartOffset); // write offset to dataStart
-
-        // from hereon there are data specific structures that are written to the IoBuffer
-        IoSerialiser<YaS, StrippedDataType>::serialise(buffer, fieldName, getAnnotatedMember(unwrapPointer(data)));
-
-        // add just data-end position
-        buffer.at<int32_t>(dataSizePosition) = static_cast<int32_t>(buffer.size() - dataStartPosition); // write data size
-
-        return dataSizePosition; // N.B. exported for adjusting START_MARKER -> END_MARKER data size to be adjustable and thus skippable
-    }
-};
-
-template<>
-inline void updateSize<YaS>(IoBuffer &buffer, const size_t posSizePositionStart, const size_t posStartDataStart) {
-    buffer.at<int32_t>(posSizePositionStart) = static_cast<int32_t>(buffer.size() - posStartDataStart); // write data size
+        callable(std::forward<decltype(member) &>(member), searchIndex);
+    });
 }
 
-template<>
-inline void putHeaderInfo<YaS>(IoBuffer &buffer) {
-    buffer.reserve_spare(2 * sizeof(int) + 7); // magic int + string length int + 4 byte 'YAS\0` string + 3 version bytes
-    buffer.put(yas::VERSION_MAGIC_NUMBER);
-    buffer.put(yas::PROTOCOL_NAME);
-    buffer.put(yas::VERSION_MAJOR);
-    buffer.put(yas::VERSION_MINOR);
-    buffer.put(yas::VERSION_MICRO);
+/**
+ * Helper function which handles an error depending on the selected type of error handling
+ * @tparam protocolCheckVariant Type of error handling
+ * @param info object which will be updated with error information
+ * @param formatString the format string for the error message
+ * @param arguments arguments for use in the format string
+ */
+inline constexpr void handleError(const ProtocolCheck protocolCheckVariant, DeserialiserInfo &info, const char *formatString, const auto &...arguments) {
+    const auto text = fmt::format(formatString, arguments...);
+    if (protocolCheckVariant == ALWAYS) {
+        throw ProtocolException(text);
+    }
+    info.exceptions.emplace_back(ProtocolException(text));
 }
 
-template<>
-inline DeserialiserInfo checkHeaderInfo<YaS>(IoBuffer &buffer, DeserialiserInfo info, const ProtocolCheck protocolCheckVariant) {
-    auto magic      = buffer.get<int>();
-    auto proto_name = buffer.get<std::string>();
-    auto ver_major  = buffer.get<int8_t>();
-    auto ver_minor  = buffer.get<int8_t>();
-    auto ver_micro  = buffer.get<int8_t>();
-    if (yas::VERSION_MAGIC_NUMBER != magic) {
-        if (protocolCheckVariant == LENIENT) {
-            info.exceptions.template emplace_back(ProtocolException(fmt::format("Wrong serialiser magic number: {} != -1", magic)));
+template<ReflectableClass T>
+std::unordered_map<std::string_view, int32_t> createMemberMap() {
+    std::unordered_map<std::string_view, int32_t> m;
+    refl::util::for_each(refl::reflect<T>().members, [&m](auto field, auto index) {
+        m.insert({ field.name.c_str(), index });
+    });
+    return m;
+}
+
+template<ReflectableClass T>
+constexpr auto createMemberMap2() noexcept {
+    constexpr size_t                                        size = refl::reflect<T>().members.size;
+    constexpr ConstExprMap<std::string_view, int32_t, size> m    = { refl::util::map_to_array<std::pair<std::string_view, int32_t>>(refl::reflect<T>().members, [](auto field, auto index) {
+        return std::pair<std::string_view, int32_t>(field.name.c_str(), index);
+    }) };
+    return m;
+}
+
+template<ReflectableClass T>
+int32_t findMemberIndex(const std::string_view fieldName) {
+    //static const std::unordered_map<std::string_view, int32_t> m = createMemberMap<T>(); // TODO: consider replacing this by ConstExprMap (array list-based)
+    static constexpr auto m = createMemberMap2<T>(); //alt: array-based implementation
+    return m.at(fieldName);
+}
+
+/**
+ * Serialises an object into an IoBuffer
+ *
+ * @tparam protocol The serialisation Type to use: YaS, Json, Yaml, CMW...
+ * @tparam writeMetaInfo whether to write metadata(unit, description) if the chosen serialiser supports it
+ * @tparam T The type of the serialised object
+ * @param buffer An IoBuffer instance
+ * @param value The object to be serialised
+ */
+template<SerialiserProtocol protocol, const bool writeMetaInfo = true, ReflectableClass T>
+constexpr void serialise(IoBuffer &buffer, const T &value) {
+    putHeaderInfo<protocol>(buffer);
+    const refl::type_descriptor<T> &reflectionData       = refl::reflect(value);
+    const auto                      type_name            = reflectionData.name.c_str();
+    std::size_t                     posSizePositionStart = FieldHeader<protocol>::template putFieldHeader<writeMetaInfo>(buffer, type_name, reflectionData.name.size, START_MARKER_INST);
+    std::size_t                     posStartDataStart    = buffer.size();
+    serialise<protocol, writeMetaInfo>(buffer, value, 0);
+    FieldHeader<protocol>::template putFieldHeader<writeMetaInfo>(buffer, type_name, reflectionData.name.size, END_MARKER_INST);
+    updateSize<protocol>(buffer, posSizePositionStart, posStartDataStart);
+}
+
+/**
+ * Helper function which serialises a sub-object at a given hierarchy depth
+ *
+ * @tparam protocol The serialisation Type to use: YaS, Json, Yaml, CMW...
+ * @tparam writeMetaInfo whether to write metadata(unit, description) if the chosen serialiser supports it
+ * @tparam T The type of the serialised object
+ * @param buffer An IoBuffer instance
+ * @param value The object to be serialised
+ * @param hierarchyDepth The level of nesting, zero is the root object
+ */
+template<SerialiserProtocol protocol, const bool writeMetaInfo = true, ReflectableClass T>
+constexpr void serialise(IoBuffer &buffer, const T &value, const uint8_t hierarchyDepth) {
+    for_each(refl::reflect(value).members, [&](const auto member, [[maybe_unused]] const auto index) {
+        if constexpr (is_field(member) && !is_static(member)) {
+            using UnwrappedMemberType = std::remove_reference_t<decltype(member(value))>;
+            using MemberType          = std::remove_reference_t<decltype(getAnnotatedMember(unwrapPointer(member(value))))>;
+            if constexpr (is_smart_pointer<std::remove_reference_t<UnwrappedMemberType>>) {
+                if (!member(value)) {
+                    return; // skip empty smart pointer
+                }
+            }
+            if constexpr (isReflectableClass<MemberType>()) { // nested data-structure
+                std::size_t posSizePositionStart = FieldHeader<protocol>::template putFieldHeader<writeMetaInfo>(buffer, member.name.c_str(), member.name.size, START_MARKER_INST);
+                std::size_t posStartDataStart    = buffer.size();
+                serialise<protocol, writeMetaInfo>(buffer, getAnnotatedMember(unwrapPointer(member(value))), hierarchyDepth + 1); // do not inspect annotation itself
+                FieldHeader<protocol>::template putFieldHeader<writeMetaInfo>(buffer, member.name.c_str(), member.name.size, END_MARKER_INST);
+                updateSize<protocol>(buffer, posSizePositionStart, posStartDataStart);
+            } else { // primitive type
+                FieldHeader<protocol>::template putFieldHeader<writeMetaInfo>(buffer, member.name.c_str(), member.name.size, member(value));
+            }
         }
-        if (protocolCheckVariant == ALWAYS) {
-            throw ProtocolException(fmt::format("Wrong serialiser magic number: {} != -1", magic));
+    });
+}
+
+template<SerialiserProtocol protocol>
+inline constexpr FieldDescription readFieldHeader(IoBuffer & /*buffer*/, DeserialiserInfo & /*info*/, const ProtocolCheck /*protocolCheckVariant*/) { return FieldDescription{}; }
+
+/**
+ * Deserialise the contents of an IoBuffer into a given object
+ * @tparam protocol The deserialisation Type to use: YaS, Json, Yaml, CMW...
+ * @tparam protocolCheckVariant determines the error and logging behaviour of the deserialiser
+ * @tparam T The type of the deserialised object
+ * @param buffer An IoBuffer instance
+ * @param value The object to be serialised
+ * @param info Object which will get populated with information about the deserialisation
+ * @return object containing info like read fields, additional fields and errors. Depends on the protocolCheck level
+ */
+template<SerialiserProtocol protocol, const ProtocolCheck protocolCheckVariant, ReflectableClass T>
+constexpr DeserialiserInfo deserialise(IoBuffer &buffer, T &value, DeserialiserInfo info = DeserialiserInfo()) {
+    // check data header for protocol version match
+    info = checkHeaderInfo<protocol>(buffer, info, protocolCheckVariant);
+    if (protocolCheckVariant == LENIENT && !info.exceptions.empty()) {
+        return info; // do not attempt to deserialise data with wrong header
+    }
+    return deserialise<protocol, protocolCheckVariant>(buffer, value, info, "root", 0);
+}
+
+/**
+ * Deserialise the object inside of an IoBuffer into an object.
+ * Internal helper function for handling nested objects.
+ * @tparam protocol The deserialisation Type to use: YaS, Json, Yaml, CMW...
+ * @tparam protocolCheckVariant determines the error and logging behaviour of the deserialiser
+ * @tparam T The type of the deserialised object
+ * @param buffer An IoBuffer instance
+ * @param value The object to be serialised
+ * @param info Object which will get populated with information about the deserialisation
+ * @param structName Name of the (sub) object
+ * @param hierarchyDepth Hierarchy level of the (sub)object. Zero means the root object.
+ * @return object containing info like read fields, additional fields and errors. Depends on the protocolCheck level
+ */
+template<SerialiserProtocol protocol, const ProtocolCheck protocolCheckVariant, ReflectableClass T>
+constexpr DeserialiserInfo deserialise(IoBuffer &buffer, T &value, DeserialiserInfo info, const std::string &structName, const uint8_t hierarchyDepth) {
+    // todo: replace structName string by const_string
+    // initialize bitfield indicating which fields have been set
+    if constexpr (protocolCheckVariant != IGNORE) {
+        if (info.setFields.contains(structName)) {
+            std::fill(info.setFields[structName].begin(), info.setFields[structName].end(), false);
+        } else {
+            info.setFields[structName] = std::vector<bool>(refl::reflect<T>().members.size);
         }
     }
-    if (yas::PROTOCOL_NAME != proto_name) {
-        if (protocolCheckVariant == LENIENT) {
-            info.exceptions.template emplace_back(ProtocolException(fmt::format("Wrong serialiser identification string: {} != YaS", proto_name)));
-        }
-        if (protocolCheckVariant == ALWAYS) {
-            throw ProtocolException(fmt::format("Wrong serialiser identification string: {} != YaS", proto_name));
-        }
+
+    // read initial field header
+    const FieldDescription startMarker = readFieldHeader<protocol>(buffer, info, protocolCheckVariant);
+    if (hierarchyDepth == 0 && startMarker.fieldName != typeName<T> && protocolCheckVariant != IGNORE) { // check if root type is matching
+        handleError(protocolCheckVariant, info, "IoSerialiser<{}, {}>::deserialise: data is not of excepted type but of type {}", protocol::protocolName(), typeName<T>, startMarker.fieldName);
     }
-    if (yas::VERSION_MAJOR != ver_major) {
-        if (protocolCheckVariant == LENIENT) {
-            info.exceptions.template emplace_back(ProtocolException(fmt::format("Major versions do not match, received {}.{}.{}", ver_major, ver_minor, ver_micro)));
+    try {
+        IoSerialiser<protocol, START_MARKER>::deserialise(buffer, startMarker.fieldName, START_MARKER_INST);
+    } catch (ProtocolException &exception) { // protocol exception
+        if constexpr (protocolCheckVariant == IGNORE) {
+            buffer.set_position(startMarker.dataEndPosition);
+            return info;
         }
-        if (protocolCheckVariant == ALWAYS) {
-            throw ProtocolException(fmt::format("Major versions do not match, received {}.{}.{}", ver_major, ver_minor, ver_micro));
+        handleError(protocolCheckVariant, info, "IoSerialiser<{}, START_MARKER>::deserialise(buffer, fieldName, START_MARKER_INST) exception for class {}: position {} vs. size {} -- exception thrown: {}",
+                protocol::protocolName(), structName, buffer.position(), buffer.size(), exception.what());
+        buffer.set_position(startMarker.dataEndPosition);
+        return info;
+    } catch (...) {
+        if constexpr (protocolCheckVariant == IGNORE) {
+            buffer.set_position(startMarker.dataEndPosition);
+            return info;
         }
+        throw ProtocolException(fmt::format("unknown exception in IoSerialiser<{}, START_MARKER>::deserialise(buffer, fieldName, START_MARKER_INST) for class {}: position {} vs. size {}",
+                protocol::protocolName(), structName, buffer.position(), buffer.size()));
     }
+    buffer.set_position(startMarker.dataStartPosition); // skip to data start
+
+    while (buffer.position() < buffer.size()) {
+        const FieldDescription field = readFieldHeader<protocol>(buffer, info, protocolCheckVariant);
+        // skip to data start
+        buffer.set_position(field.dataStartPosition);
+
+        if (field.intDataType == IoSerialiser<protocol, END_MARKER>::getDataTypeId()) {
+            // todo: assert name equals start marker
+            // reached end of sub-structure
+            try {
+                IoSerialiser<protocol, END_MARKER>::deserialise(buffer, field.fieldName, END_MARKER_INST);
+            } catch (ProtocolException &exception) { // protocol exception
+                if constexpr (protocolCheckVariant == IGNORE) {
+                    buffer.set_position(field.dataEndPosition);
+                    continue;
+                }
+                const auto text = fmt::format("IoSerialiser<{}, END_MARKER>::deserialise(buffer, fieldName, END_MARKER_INST) exception for class {}: position {} vs. size {} -- exception thrown: {}",
+                        protocol::protocolName(), structName, buffer.position(), buffer.size(), exception.what());
+                if constexpr (protocolCheckVariant == ALWAYS) {
+                    throw ProtocolException(text);
+                }
+                info.exceptions.emplace_back(ProtocolException(text));
+                buffer.set_position(field.dataEndPosition);
+                continue;
+            } catch (...) {
+                if constexpr (protocolCheckVariant == IGNORE) {
+                    buffer.set_position(field.dataEndPosition);
+                    continue;
+                }
+                throw ProtocolException(fmt::format("unknown exception in IoSerialiser<{}, START_MARKER>::deserialise(buffer, fieldName, START_MARKER_INST) for class {}: position {} vs. size {}",
+                        protocol::protocolName(), structName, buffer.position(), buffer.size()));
+            }
+            return info; // step down to previous hierarchy depth
+        }
+
+        int32_t searchIndex = -1;
+        try {
+            searchIndex = static_cast<int32_t>(findMemberIndex<T>(field.fieldName));
+        } catch (std::out_of_range &e) {
+            if constexpr (protocolCheckVariant == IGNORE) {
+                buffer.set_position(field.dataEndPosition);
+                continue;
+            }
+            const auto exception = fmt::format("missing field (type:{}) {}::{} at buffer[{}, size:{}]",
+                    field.intDataType, structName, field.fieldName, buffer.position(), buffer.size());
+            if constexpr (protocolCheckVariant == ALWAYS) {
+                throw ProtocolException(exception);
+            }
+            info.exceptions.emplace_back(ProtocolException(exception));
+            info.additionalFields.emplace_back(std::make_tuple(fmt::format("{}::{}", structName, field.fieldName), field.intDataType));
+            buffer.set_position(field.dataEndPosition);
+            continue;
+        }
+        if (field.intDataType == IoSerialiser<protocol, START_MARKER>::getDataTypeId()) {
+            buffer.set_position(field.headerStart); // reset buffer position for the nested deserialiser to read again
+            // reached start of sub-structure -> dive in
+            call_on_field<T>(field.fieldName, [&buffer, &value, &hierarchyDepth, &info, &structName](auto member, int32_t index) {
+                using MemberType = std::remove_reference_t<decltype(getAnnotatedMember(unwrapPointer(member(value))))>;
+                if constexpr (isReflectableClass<MemberType>()) {
+                    info = deserialise<protocol, protocolCheckVariant>(buffer, unwrapPointerCreateIfAbsent(member(value)), info, fmt::format("{}.{}", structName, get_display_name(member)), hierarchyDepth + 1);
+                    if constexpr (protocolCheckVariant != IGNORE) {
+                        info.setFields[structName][static_cast<uint64_t>(index)] = true;
+                    }
+                }
+            });
+
+            buffer.set_position(field.dataEndPosition);
+            continue;
+        }
+
+        call_on_field<T>(field.fieldName, [&buffer, &value, &info, &field, &searchIndex, &structName](auto member, int32_t index) {
+            using MemberType = std::remove_reference_t<decltype(getAnnotatedMember(unwrapPointer(member(value))))>;
+            if constexpr (isReflectableClass<MemberType>() || !is_writable(member) || is_static(member)) {
+                handleError(protocolCheckVariant, info, "field is not writeable or non-primitive: {}", member.name);
+                return;
+            } else {
+                constexpr int requestedType = IoSerialiser<protocol, MemberType>::getDataTypeId();
+                if (requestedType != field.intDataType) { // mismatching data-type
+                    if constexpr (protocolCheckVariant == IGNORE) {
+                        return; // don't write -> skip to next
+                    }
+                    handleError(protocolCheckVariant, info, "mismatched field type for {}::{} - requested type: {} (typeID: {}) got: {}", member.declarator.name, member.name, typeName<MemberType>, requestedType, field.intDataType);
+                    return;
+                }
+                constexpr bool isAnnotated = is_annotated<std::remove_reference_t<decltype(unwrapPointer(member(value)))>>;
+                if constexpr (isAnnotated && protocolCheckVariant != IGNORE) {       // check for Annotation mismatch
+                    if (is_deprecated(unwrapPointer(member(value)).getModifier())) { // warn for deprecated access
+                        handleError(protocolCheckVariant, info, "deprecated field access for {}::{} - description: {}", member.declarator.name, member.name, unwrapPointer(member(value)).getDescription());
+                    }
+                    if (is_private(unwrapPointer(member(value)).getModifier())) { // warn for private access
+                        handleError(protocolCheckVariant, info, "private/internal field access for {}::{} - description: {}", member.declarator.name, member.name, unwrapPointer(member(value)).getDescription());
+                    }
+                    if (unwrapPointer(member(value)).getUnit().compare(field.unit)) { // error on unit mismatch
+                        handleError(protocolCheckVariant, info, "mismatched field unit for {}::{} - requested unit '{}' received '{}'", member.declarator.name, member.name, unwrapPointer(member(value)).getUnit(), field.unit);
+                        return;
+                    }
+                    if (is_readonly((unwrapPointer(member(value)).getModifier()))) { // should not set field via external reference
+                        handleError(protocolCheckVariant, info, "mismatched field access modifier for {}::{} - requested '{}' received '{}'", member.declarator.name, member.name, (unwrapPointer(member(value)).getModifier()), field.modifier);
+                        return;
+                    }
+                }
+                IoSerialiser<protocol, MemberType>::deserialise(buffer, field.fieldName, getAnnotatedMember(unwrapPointerCreateIfAbsent(member(value))));
+                if constexpr (protocolCheckVariant != IGNORE) {
+                    info.setFields[structName][static_cast<uint64_t>(index)] = true;
+                }
+            }
+        });
+        // skip to data end
+        buffer.set_position(field.dataEndPosition);
+    }
+    // check that full buffer is read. // todo: this check should be removed to allow consecutive storage of multiple objects in one buffer
+    if (hierarchyDepth == 0 && buffer.position() != buffer.size()) {
+        if constexpr (protocolCheckVariant == IGNORE) {
+            return info;
+        }
+        std::cerr << "serialise class type " << typeName<T> << " hierarchyDepth = " << static_cast<int>(hierarchyDepth) << '\n';
+        const auto exception = fmt::format("protocol exception for class type {}({}): position {} vs. size {}",
+                typeName<T>, static_cast<int>(hierarchyDepth), buffer.position(), buffer.size());
+        if constexpr (protocolCheckVariant == ALWAYS) {
+            throw ProtocolException(exception);
+        }
+        std::cout << exception << std::endl; //TODO: replace std::cerr/cout by logger?
+        info.exceptions.emplace_back(ProtocolException(exception));
+        return info;
+    }
+
     return info;
 }
-
 } // namespace opencmw
-
-/* #################################################################################### */
-/* ### Json Definitions ########################################################### */
-/* #################################################################################### */
-namespace opencmw {
-
-struct Json : Protocol<"Json"> {}; // todo: move to more appropriate place, but has to be declared before usage
-
-namespace json {
-}
-
-template<>
-struct FieldHeader<Json> {
-    template<const bool writeMetaInfo, typename DataType>
-    constexpr std::size_t static putFieldHeader(IoBuffer &buffer, const char *fieldName, const int fieldNameSize, const DataType &data) { // todo fieldName -> string_view
-        if constexpr (std::is_same_v<DataType, START_MARKER>) {
-            if (fieldNameSize > 0) {
-                buffer.putRaw(fieldName);
-                buffer.putRaw(":");
-            }
-            buffer.putRaw("{\n"); // use string put instead of other put // call serialise with the start marker instead?
-            return 0;
-        }
-        if constexpr (std::is_same_v<DataType, END_MARKER>) {
-            buffer.put('}');
-            return 0;
-        }
-        buffer.putRaw(std::basic_string_view(fieldName, static_cast<size_t>(fieldNameSize)));
-        buffer.put(':');
-        using StrippedDataType = std::remove_reference_t<decltype(getAnnotatedMember(unwrapPointer(data)))>;
-        IoSerialiser<Json, StrippedDataType>::serialise(buffer, fieldName, getAnnotatedMember(unwrapPointer(data)));
-        buffer.putRaw(",\n");
-        return 0; // return value is irrelevant for Json
-    }
-};
-
-template<>
-struct IoSerialiser<Json, END_MARKER> { // catch all template
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const END_MARKER & /*value*/) noexcept {
-        buffer.putRaw("}");
-        return false;
-    }
-    constexpr static bool deserialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const END_MARKER &) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-};
-
-template<>
-struct IoSerialiser<Json, START_MARKER> { // catch all template
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const START_MARKER & /*value*/) noexcept {
-        buffer.putRaw("{");
-        return false;
-    }
-    constexpr static bool deserialise(IoBuffer & /*buffer*/, const ClassField & /*field*/, const START_MARKER & /*value*/) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        return std::is_constant_evaluated();
-    }
-};
-
-inline bool isJsonNumberChar(uint8_t c) {
-    switch (c) {
-    case '-':
-    case '+':
-    case '1':
-    case '2':
-    case '3':
-    case '4':
-    case '5':
-    case '6':
-    case '7':
-    case '8':
-    case '9':
-    case '0':
-    case '.':
-    case 'e':
-    case 'E':
-        return true;
-    default:
-        return false;
-    }
-}
-
-template<Number T>
-struct IoSerialiser<Json, T> { // catch all template
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    /*constexpr*/ static bool       serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        // todo: constexpr not possible because of fmt
-        if constexpr (std::is_integral_v<T>) {
-            // buffer.putRaw(std::to_string(value));
-            buffer.putRaw(fmt::format("{}", value));
-        } else {
-            buffer.putRaw(fmt::format("{}", value));
-        }
-        return false;
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        auto start = buffer.position();
-        while (isJsonNumberChar(buffer.get<uint8_t>())) {
-        }
-        auto end = buffer.position();
-        // value = stringTo
-        std::ignore = start;
-        std::ignore = end;
-        std::ignore = value;
-        return std::is_constant_evaluated();
-    }
-};
-template<StringLike T>
-struct IoSerialiser<Json, T> { // catch all template
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        buffer.put('"');
-        buffer.putRaw(value);
-        buffer.put('"');
-        return false;
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        auto start = buffer.position();
-        while (isJsonNumberChar(buffer.get<uint8_t>())) {
-        }
-        auto end = buffer.position();
-        // value = stringTo
-        std::ignore = start;
-        std::ignore = end;
-        std::ignore = value;
-        return std::is_constant_evaluated();
-    }
-};
-
-template<ArrayOrVector T>
-struct IoSerialiser<Json, T> {
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &values) noexcept {
-        using MemberType = typename T::value_type;
-        buffer.put('[');
-        for (auto value : values) {
-            // todo: use same formatting as for non array elements
-            IoSerialiser<Json, MemberType>::serialise(buffer, "", value);
-            buffer.putRaw(", ");
-        }
-        buffer.resize(buffer.size() - 2); // remove trailing comma
-        buffer.put(']');
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        auto start = buffer.position();
-        while (isJsonNumberChar(buffer.get<uint8_t>())) {
-        }
-        auto end = buffer.position();
-        // value = stringTo
-        std::ignore = start;
-        std::ignore = end;
-        std::ignore = value;
-        return std::is_constant_evaluated();
-    }
-};
-
-template<MultiArrayType T>
-struct IoSerialiser<Json, T> {
-    inline static constexpr uint8_t getDataTypeId() { return 0; }
-    constexpr static bool           serialise(IoBuffer &buffer, const ClassField & /*field*/, const T &value) noexcept {
-        buffer.putRaw("{\n");
-        std::array<int32_t, T::n_dims_> dims;
-        for (uint32_t i = 0U; i < T::n_dims_; i++) {
-            dims[i] = static_cast<int32_t>(value.dimensions()[i]);
-        }
-        FieldHeader<Json>::template putFieldHeader<false>(buffer, "dims", 4, dims);
-        FieldHeader<Json>::template putFieldHeader<false>(buffer, "values", 6, value.elements());
-        buffer.putRaw("}");
-        return std::is_constant_evaluated();
-    }
-    constexpr static bool deserialise(IoBuffer &buffer, const ClassField & /*field*/, T &value) {
-        // do not do anything, as the end marker is of size zero and only the type byte is important
-        auto start = buffer.position();
-        while (isJsonNumberChar(buffer.get<uint8_t>())) {
-        }
-        auto end = buffer.position();
-        // value = stringTo
-        std::ignore = start;
-        std::ignore = end;
-        std::ignore = value;
-        return std::is_constant_evaluated();
-    }
-};
-
-} // namespace opencmw
-
-/* #################################################################################### */
-/* ### CmwLight Definitions ########################################################### */
-/* #################################################################################### */
-
-namespace opencmw {
-struct CmwLight : Protocol<"CmwLight"> {
-};
-
-namespace cmwlight {
-
-}
-
-template<Number T>
-struct IoSerialiser<T, CmwLight> { // catch all template
-    constexpr static bool serialise(IoBuffer &buffer, const ClassField &field, const T &value) noexcept {
-        if (std::is_constant_evaluated()) {
-            using namespace std::literals;
-            buffer.put(field);
-            buffer.template put(value);
-            return true;
-        }
-        std::cout << fmt::format("{} - serialise-catch-all: {} {} value: {} - constexpr?: {}\n",
-                CmwLight::protocolName(), typeName<T>, field, value,
-                std::is_constant_evaluated());
-        return false;
-    }
-};
-} // namespace opencmw
-
 // TODO: allow declaration of reflection within name-space
 ENABLE_REFLECTION_FOR(opencmw::DeserialiserInfo, setFields, additionalFields, exceptions)
