+++ src/serialiser/test/IoClassSerialiser_tests.cpp
@@ -3,9 +3,8 @@
 #pragma ide diagnostic   ignored "cppcoreguidelines-avoid-magic-numbers"
 
 #include <Debug.hpp>
-#include <IoClassSerialiser.hpp>
-#include <IoSerialiser.hpp>
 #include <Utils.hpp>
+#include <YaSerialiser.hpp>
 #include <algorithm>
 #include <catch2/catch.hpp>
 #include <iostream>
